+++ photons_app/__init__.py
@@ -1,4 +1,4 @@
-VERSION = "0.9.2"
+VERSION = "0.9.3"
 
 __shortdesc__ = """Base module for all photons applications"""