+++ docs/start/changelog.rst
@@ -6,6 +6,7 @@ ChangeLog
 0.9.3 - TBD
     * Minor changes
     * Another efficiency improvement for tile animations
+    * Some fixes to the scrolling animations
 
 0.9.2 - 27 December 2018
     * Made tile_marquee work without options